requirements.txt
blinker==1.9.0
click==8.1.8
docopt==0.6.2
Flask==3.1.1
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
paho-mqtt==2.1.0
python-dotenv==1.1.1
stomp.py==8.2.0
websocket-client==1.8.0
Werkzeug==3.1.3
zipp==3.23.0
